Back-End/requirements.txt
alembic==1.6.5
appnope==0.1.2
astroid==2.5.1
attrs==20.3.0
backcall==0.2.0
certifi==2020.12.5
chardet==4.0.0
click==8.0.1
decorator==5.0.5
Flask==2.0.1
Flask-Cors==3.0.10
Flask-Migrate==3.0.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
gunicorn==20.1.0
idna==2.10
iniconfig==1.1.1
ipykernel==5.5.3
ipython==7.22.0
ipython-genutils==0.2.0
isort==5.7.0
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
jupyter-client==6.1.12
jupyter-core==4.7.1
lazy-object-proxy==1.5.2
Mako==1.1.4
MarkupSafe==2.0.1
mccabe==0.6.1
packaging==20.9
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
pluggy==0.13.1
prompt-toolkit==3.0.18
psycopg2==2.9.1
ptyprocess==0.7.0
py==1.10.0
pygame==2.0.1
Pygments==2.8.1
pylint==2.7.2
pyparsing==2.4.7
pytest==6.2.2
python-dateutil==2.8.1
python-dotenv==0.18.0
python-editor==1.0.4
pyzmq==22.0.3
requests==2.25.1
six==1.15.0
SQLAlchemy==1.3.24
PyMySQL==1.0.2
toml==0.10.2
tornado==6.1
traitlets==5.0.5
urllib3==1.26.4
wcwidth==0.2.5
Werkzeug==2.0.1
wonderwords==2.2.0
wrapt==1.12.1
